free grid 2

+++ 0x0B-malloc_free/4-free_grid.c
@@ -11,7 +11,7 @@
  */
 void free_grid(int **grid, int height)
 {
-	size_t i, j, width;
+	int i, j, width;
 
 	for (i = 0; i < height; i++)
 	{
